requirements.txt
PyMuPDF==1.26.3
python-docx==1.2.0
extract-msg==0.54.1
